+++ bin/metrics.dart
@@ -6,7 +6,6 @@ import 'package:dart_code_metrics/src/cli/arguments_parser.dart';
 import 'package:dart_code_metrics/src/cli/arguments_validation.dart';
 import 'package:dart_code_metrics/src/cli/arguments_validation_exceptions.dart';
 import 'package:dart_code_metrics/src/models/violation_level.dart';
-import 'package:dart_code_metrics/src/reporters/github/github_reporter.dart';
 import 'package:dart_code_metrics/src/reporters/utility_selector.dart';
 import 'package:path/path.dart' as p;
 
@@ -75,8 +74,8 @@ Future<void> _runAnalysis(
   final store = MetricsRecordsStore.store();
   final analyzer = MetricsAnalyzer(store,
       options: options, addintionalExcludes: [ignoreFilesPattern]);
-  final runner = MetricsAnalysisRunner(analyzer, store, analysisDirectories,
-      rootFolder: rootFolder);
+  final runner =
+      MetricsAnalysisRunner(analyzer, store, analysisDirectories, rootFolder);
   await runner.run();
 
   final config = Config(