+++ lib/src/rules_factory.dart
@@ -7,9 +7,11 @@ import 'rules/member_ordering.dart';
 import 'rules/newline_before_return.dart';
 import 'rules/no_boolean_literal_compare_rule.dart';
 import 'rules/no_empty_block.dart';
+import 'rules/no_equal_arguments.dart';
 import 'rules/no_equal_then_else.dart';
 import 'rules/no_magic_number_rule.dart';
 import 'rules/no_object_declaration.dart';
+import 'rules/potential_null_dereference.dart';
 import 'rules/prefer_conditional_expressions.dart';
 import 'rules/prefer_intl_name.dart';
 import 'rules/prefer_on_push_cd_strategy.dart';
@@ -31,10 +33,13 @@ final _implementedRules = <String, BaseRule Function(Map<String, Object>)>{
   NoBooleanLiteralCompareRule.ruleId: (config) =>
       NoBooleanLiteralCompareRule(config: config),
   NoEmptyBlockRule.ruleId: (config) => NoEmptyBlockRule(config: config),
+  NoEqualArguments.ruleId: (config) => NoEqualArguments(config: config),
   NoEqualThenElse.ruleId: (config) => NoEqualThenElse(config: config),
   NoMagicNumberRule.ruleId: (config) => NoMagicNumberRule(config: config),
   NoObjectDeclarationRule.ruleId: (config) =>
       NoObjectDeclarationRule(config: config),
+  PotentialNullDereference.ruleId: (config) =>
+      PotentialNullDereference(config: config),
   PreferConditionalExpressions.ruleId: (config) =>
       PreferConditionalExpressions(config: config),
   ProvideCorrectIntlArgsRule.ruleId: (config) =>